pyproject.toml
[project]
name = "ams-chatbot"
version = "0.1.0"
description = "A Flask-based microservice for the Airline Management System chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.5",
    "jq>=1.8.0",
    "langchain-google-genai>=2.0.10",
    "mcp[cli]>=1.7.1",
    "psycopg",
    "psycopg2",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "gunicorn>=23.0.0",
]
